test: share one MuxSettingsView fixture, reclaim 14 rotted tests (#304)
First cluster of #304. Fixes 14 of the 41 failures in the excluded root
suite, and adds the first reclaimed file back to the CI test scope.

## Prior art: this exact rot was fixed once and lost

be4688a ("test(observability): Phase A foundation coverage + un-break stale
mux helpers", 25 May) fixed the same fields in the same files seven weeks
ago: "fixes stale `makeSettingsView` helpers (missing cache / metricsEnabled
fields start() reads unconditionally) ... repairs 7 pre-existing failures".

It is not in main — `git merge-base --is-ancestor be4688a povai/main` says
NOT ANCESTOR, and main carries that commit's parent blob. The fix was dropped
during an upstream-sync graft. a99c527 is a third instance of patching this
same fixture in place.

Three local repairs, two of them lost or re-rotted. That is why this fixes it
structurally rather than in place.

## The rot

Three copies of `makeSettingsView` existed:

  src/plugins/mcp-multiplexer/__tests__/index.test.ts   <- adjacent to the code
  src/__tests__/mcp-multiplexer-integration.test.ts     <- orphaned
  src/__tests__/session-persistence-integration.test.ts <- orphaned

When #68 added `metricsEnabled` (931b25a) and #69 added `cache` (cde2bd8) as
REQUIRED fields on `MuxSettingsView`, both commits updated the copy sitting
next to the code they changed and neither touched the two under
`src/__tests__/`. Nothing ran any of them to catch it — there was no CI test
job at all until 7a0db7f, six weeks later, and it has never covered these
paths. The two orphans kept building an object missing both fields, so every
test in them died at `src/plugins/mcp-multiplexer/index.ts:296`:

    TypeError: undefined is not an object (evaluating 'settings.cache.cacheable')

0 pass / 14 fail across the two files. `tsc --noEmit` flagged the stale
fixture the whole time (TS2322 at mcp-multiplexer-integration.test.ts:65) —
there is no typecheck step to surface it (#327).

This is #304's own thesis demonstrated: excluding tests from CI is what let
them rot.

## The fix

One shared builder at `src/__tests__/fixtures/mux-settings-view.ts`, used by
all THREE files — including the surviving in-plugin copy. Consolidating only
the two broken ones would have left the same rot vector in place. Precedent
for the import direction already existed: the plugin test imports
`mock-mcp-server.ts` from the same fixtures dir.

The next required field now breaks every call site at once, loudly, instead
of only the ones nobody runs. Verified the three copies' defaults were
byte-identical before folding them together.

Defaults match the surviving copy (timers off, optional subsystems off) so a
test opts in to what it exercises. All 8 persistence call sites already set
`sessionPersistenceEnabled` explicitly, so the shared default does not change
their behaviour — checked rather than assumed.

## Scope: fixture consolidation only

An earlier revision of this branch also converted the persistence file's audit
lookups to awaited polling, on the theory that a bare `events.find()` raced an
async audit. Review disproved that: `_audit` is a direct synchronous call and
the test's capture wrapper pushes synchronously, so both events are in the
array before the awaited call returns. There was no race. That machinery is
reverted; this PR touches only the fixtures.

The persistence file's remaining flake is a wrong PAYLOAD, not a missing
event (`scanned: 1` where 2 entries were verifiably on disk a line earlier) —
i.e. something rewrites the file between the test's append and the GC pass.
Correct diagnosis recorded on #326, replacing the wrong one I filed there.

## biome: the fixtures dir was exempt from lint and format

`biome.json` excluded all of `src/__tests__/fixtures`, presumably aimed at
`mock-mcp-server.ts` (a standalone script spawned as a child). That dir now
holds shared test infrastructure with call sites across three suites, and a
shared file nothing formats or lints is a poor foundation — the new fixture
had in fact never been formatted.

Narrowed the exclusion to the mock script itself.

## CI scope

Adds `src/__tests__/mcp-multiplexer-integration.test.ts` only: 7 pass,
verified 6/6 across repeat runs, no known flake.

`session-persistence-integration.test.ts` is fixed (7 pass) but deliberately
held back until #326 lands — adding a test that fails ~25% under load would
make CI randomly red, which is the exact failure mode #304 exists to remove.

## Verification

  mcp-multiplexer     0 pass /  7 fail  ->   7 pass / 0 fail
  session-persistence 0 pass /  7 fail  ->   6-7 pass (1 pre-existing flake, #326)
  surviving 3rd copy  29 pass / 0 fail   (unchanged, now on the shared fixture)
  all three together  43 pass / 0 fail
  full CI scope       1364 pass / 0 fail

The one failure in the widened-scope run was `session-id collision rotation`
in src/bus — a pre-existing load-sensitive flake, already in CI scope, not
touched by this change. Noted on #326.
